extracted balances xt's to submodule
